test(expression): move the [expression] tests into their own TU (CoolProp-u7wz)
Review feedback on #3185: the DSL's 636-line test block does not belong bolted
onto the end of CoolProp-Tests.cpp.  Moved to
src/Tests/CoolProp-Tests-Expression.cpp, following the existing
CoolProp-Tests-<Topic>.cpp convention, with the four scattered mid-file
(the block was written in the same unformatted hand style as the rest of the
module -- see the preceding style commit).

CoolProp-Tests.cpp is now byte-identical to its pre-branch state.
No test content changed: 254 assertions in 26 cases, all green.
